SeismicQ.jl

Tout ce que vous avez toujours voulu savoir sur le Q.

Package Features

Compute source functions

Function Documentation

SeismicQ.Ricker — Function

Ricker(t, t₀, 𝑓₀)

Compute the Ricker function for t, t₀ and 𝑓₀.

\[f = (1.0 - 2.0 (\pi 𝑓₀ (t - t₀))^2) \exp(-π^2 𝑓₀^2 (t - t₀)^2)\]

Examples

julia>  Ricker(0.0, 0.0, 0.0)
1.0

SeismicQ.Spec — Function

Spec(trace,t0,dt,Nt,tp,ts,\Deltaph)

Returns the amplitude spectrums of the P ans S phases where amplitudes are significant

    trace: the trace recorded at a given virtual station, starting at -t0, with time sampling dt, Nt samples
    tp: picked P-wave phase (s)
    ts: picked S-wave phase (s)
    \Deltaph: width of the phases (s) 

    Vec2dP: modulus of the fft coeffs of P phase where larger than max(Vec2dp)/10 
    Vec2dS: modulus of the fft coeffs of S phase where larger than max(Vec2dp)/10
    FrequP: corresponding frequency vector to plot P-wave amp. spectrum
    FrequS: corresponding frequency vector to plot S-wave amp. spectrum

Examples

julia>  (AmpP,AmpS,fP,fS)=Spec(trace1,0.1,1e-3,2000,0.7,1.25,0.35))
Trace length is 2.0 s using 2000 samples
Trace 1 at offset 3000 m is found at index 30
Trace 2 at offset 5000 m is found at index 50
Trace 1:
phase P is picked at 0.4 s
phase S is picked at 0.75 s
phase width is set to 0.35
Trace 2:
phase P is picked at 0.7 s
phase S is picked at 1.25 s
phase width is set to 0.35
+ Figure with 8 subplots

SeismicQ.ComputeQgraph — Function

ComputeQgraph(amp1,amp2,freq,index,d1,d2,V)

Returns x and y to plot Q-graph for a given phase using the results of Spec() for two different traces from two stations

    Can't write it for now (no LateX in online Editor?)

Examples

julia>  TBD

SeismicQ.Getfreq — Function

Getfreq(dt,Nt)

Returns the frequency vector corresponding to the fft results

    dt is the time sample in s
    Nt is the number of samples in the trace given to fft()
    The frequency increment is 1/(Nt*dt)

Examples

julia>  Getfreq(1e-3,2000)
2000-element Vector{Float64}:
    0.5
    1.0
    1.5
    2.0
    2.5
    3.0
    3.5
    4.0
    ⋮
  996.5
  997.0
  997.5
  998.0
  998.5
  999.0
  999.5
 1000.0